backend/gan_models.py
"""
Module for integrating actual Cyclic GAN models.
This is a placeholder showing how real model integration would work.
"""
import os
import logging
from typing import Optional, Dict, Any, Tuple
import time
from PIL import Image
import numpy as np

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Placeholder for real model imports
# import torch
# import tensorflow as tf
# from models.cycle_gan import Generator

class GanModelBase:
    """Base class for GAN model integration"""
    
    def __init__(self, model_path: str, device: str = "cpu"):
        """
        Initialize the GAN model
        
        Args:
            model_path: Path to model weights
            device: Device to run inference on ('cpu' or 'cuda')
        """
        self.model_path = model_path
        self.device = device
        self.model = None
        self._load_model()
        
    def _load_model(self) -> None:
        """Load the model from disk - to be implemented by subclasses"""
        raise NotImplementedError
        
    def preprocess(self, image: Image.Image) -> np.ndarray:
        """
        Preprocess the image for model input
        
        Args:
            image: PIL Image to preprocess
            
        Returns:
            Preprocessed image as numpy array
        """
        raise NotImplementedError
        
    def postprocess(self, output: np.ndarray) -> Image.Image:
        """
        Postprocess model output to PIL Image
        
        Args:
            output: Model output as numpy array
            
        Returns:
            Processed PIL Image
        """
        raise NotImplementedError
        
    def generate(self, input_path: str, output_path: str) -> bool:
        """
        Generate transformed image
        
        Args:
            input_path: Path to input image
            output_path: Path to save output image
            
        Returns:
            Success flag
        """
        try:
            # Load the image
            image = Image.open(input_path)
            
            # Preprocess
            model_input = self.preprocess(image)
            
            # Run inference
            model_output = self._run_inference(model_input)
            
            # Postprocess
            result_image = self.postprocess(model_output)
            
            # Save the result
            result_image.save(output_path)
            
            return True
        except Exception as e:
            logger.error(f"Error generating image: {e}")
            return False
            
    def _run_inference(self, input_data: np.ndarray) -> np.ndarray:
        """
        Run model inference
        
        Args:
            input_data: Preprocessed input data
            
        Returns:
            Model output
        """
        raise NotImplementedError


class StyleTransferGANIntegration(GanModelBase):
    """Integration for Style Transfer GAN"""
    
    def _load_model(self) -> None:
        """Load style transfer model"""
        logger.info(f"Loading style transfer model from {self.model_path}")
        
        # In a real implementation:
        # self.model = Generator()
        # self.model.load_state_dict(torch.load(self.model_path))
        # self.model.to(self.device)
        # self.model.eval()
        
        # For this mock version, we'll just simulate loading
        time.sleep(1)
        logger.info("Style transfer model loaded successfully")
        
    def preprocess(self, image: Image.Image) -> np.ndarray:
        """Preprocess for style transfer"""
        # Resize to model input size
        image = image.resize((256, 256))
        
        # Convert to numpy and normalize
        img_array = np.array(image).astype('float32')
        img_array = img_array / 127.5 - 1  # Normalize to [-1, 1]
        
        # Add batch dimension
        img_array = np.expand_dims(img_array, axis=0)
        
        return img_array
        
    def postprocess(self, output: np.ndarray) -> Image.Image:
        """Postprocess style transfer output"""
        # Remove batch dimension
        output = output[0]
        
        # Denormalize
        output = (output + 1) * 127.5
        output = np.clip(output, 0, 255).astype(np.uint8)
        
        # Convert to PIL Image
        return Image.fromarray(output)
        
    def _run_inference(self, input_data: np.ndarray) -> np.ndarray:
        """Run style transfer inference"""
        logger.info("Running style transfer inference")
        
        # In a real implementation:
        # with torch.no_grad():
        #     input_tensor = torch.tensor(input_data).to(self.device)
        #     output_tensor = self.model(input_tensor)
        #     return output_tensor.cpu().numpy()
        
        # For this mock version, just apply a filter to simulate style transfer
        time.sleep(2)  # Simulate processing time
        
        # Apply sepia filter to input data for demonstration
        img = input_data[0]  # Remove batch dimension
        r, g, b = img[:, :, 0], img[:, :, 1], img[:, :, 2]
        sepia_r = 0.393 * r + 0.769 * g + 0.189 * b
        sepia_g = 0.349 * r + 0.686 * g + 0.168 * b
        sepia_b = 0.272 * r + 0.534 * g + 0.131 * b
        
        sepia_img = np.stack([sepia_r, sepia_g, sepia_b], axis=-1)
        sepia_img = np.clip(sepia_img, -1, 1)
        
        # Add batch dimension back
        return np.expand_dims(sepia_img, axis=0)


class PhotoEnhancementGANIntegration(GanModelBase):
    """Integration for Photo Enhancement GAN"""
    
    def _load_model(self) -> None:
        """Load photo enhancement model"""
        logger.info(f"Loading photo enhancement model from {self.model_path}")
        
        # In a real implementation:
        # self.model = Generator()
        # self.model.load_state_dict(torch.load(self.model_path))
        # self.model.to(self.device)
        # self.model.eval()
        
        # For this mock version, we'll just simulate loading
        time.sleep(1)
        logger.info("Photo enhancement model loaded successfully")
        
    def preprocess(self, image: Image.Image) -> np.ndarray:
        """Preprocess for photo enhancement"""
        # Resize to model input size
        image = image.resize((512, 512))
        
        # Convert to numpy and normalize
        img_array = np.array(image).astype('float32')
        img_array = img_array / 255.0  # Normalize to [0, 1]
        
        # Add batch dimension
        img_array = np.expand_dims(img_array, axis=0)
        
        return img_array
        
    def postprocess(self, output: np.ndarray) -> Image.Image:
        """Postprocess photo enhancement output"""
        # Remove batch dimension
        output = output[0]
        
        # Denormalize
        output = output * 255.0
        output = np.clip(output, 0, 255).astype(np.uint8)
        
        # Convert to PIL Image
        return Image.fromarray(output)
        
    def _run_inference(self, input_data: np.ndarray) -> np.ndarray:
        """Run photo enhancement inference"""
        logger.info("Running photo enhancement inference")
        
        # In a real implementation:
        # with torch.no_grad():
        #     input_tensor = torch.tensor(input_data).to(self.device)
        #     output_tensor = self.model(input_tensor)
        #     return output_tensor.cpu().numpy()
        
        # For this mock version, just apply a filter to simulate enhancement
        time.sleep(2)  # Simulate processing time
        
        # Convert to grayscale and back for demonstration
        img = input_data[0]  # Remove batch dimension
        grayscale = np.mean(img, axis=-1, keepdims=True)
        grayscale = np.repeat(grayscale, 3, axis=-1)
        
        # Enhance contrast a bit
        grayscale = (grayscale - 0.5) * 1.2 + 0.5
        grayscale = np.clip(grayscale, 0, 1)
        
        # Add batch dimension back
        return np.expand_dims(grayscale, axis=0)


# Factory function to get the appropriate GAN model integration
def get_gan_model(task_id: str, device: str = "cpu") -> Optional[GanModelBase]:
    """
    Get the appropriate GAN model based on task ID
    
    Args:
        task_id: Task identifier
        device: Inference device
        
    Returns:
        Appropriate GAN model integration instance
    """
    # Model paths would typically come from config
    models = {
        "task1": StyleTransferGANIntegration("models/style_transfer.pth", device),
        "task2": PhotoEnhancementGANIntegration("models/photo_enhancement.pth", device)
    }
    
    return models.get(task_id)