backend/models.py
from PIL import Image
import time
import os

class BaseGANModel:
    """Base class for GAN models"""
    def __init__(self, name):
        self.name = name
    
    def generate(self, input_path, output_path):
        """
        Generate an image transformation
        
        Args:
            input_path (str): Path to input image
            output_path (str): Path to save output image
            
        Returns:
            bool: Success or failure
        """
        raise NotImplementedError("Subclasses must implement the generate method")

# --------- yo guys place task 1 model here ---------
class StyleTransferGAN(BaseGANModel):
    """Mock Style Transfer GAN implementation"""
    def __init__(self):
        super().__init__("style_transfer")
        print(f"Initializing {self.name} model...")
        # In a real implementation, you would load your model weights here
    
    def generate(self, input_path, output_path):
        """Apply style transfer to the input image"""
        try:
            # Simulate processing time
            time.sleep(2)
            
            # Load image
            img = Image.open(input_path)
            
            # Mock style transfer with a sepia filter
            width, height = img.size
            pixels = img.load()
            for i in range(width):
                for j in range(height):
                    r, g, b = pixels[i, j][:3]
                    tr = int(0.393 * r + 0.769 * g + 0.189 * b)
                    tg = int(0.349 * r + 0.686 * g + 0.168 * b)
                    tb = int(0.272 * r + 0.534 * g + 0.131 * b)
                    pixels[i, j] = (min(tr, 255), min(tg, 255), min(tb, 255))
                    
            # Save the result
            img.save(output_path)
            return True
        except Exception as e:
            print(f"Error in style transfer: {e}")
            return False

# --------- yo guys place task 2 model here ---------
class PhotoEnhancementGAN(BaseGANModel):
    """Mock Photo Enhancement GAN implementation"""
    def __init__(self):
        super().__init__("photo_enhancement")
        print(f"Initializing {self.name} model...")
        # In a real implementation, you would load your model weights here
    
    def generate(self, input_path, output_path):
        """Enhance the input photo"""
        try:
            # Simulate processing time
            time.sleep(2)
            
            # Load image
            img = Image.open(input_path)
            
            # Mock enhancement with a simple contrast adjustment
            # Convert to grayscale for this mock
            img = img.convert('L')
            
            # Enhance contrast a bit for demonstration
            from PIL import ImageEnhance
            enhancer = ImageEnhance.Contrast(img)
            img = enhancer.enhance(1.5)
            
            # Convert back to RGB
            img = img.convert('RGB')
            
            # Save the result
            img.save(output_path)
            return True
        except Exception as e:
            print(f"Error in photo enhancement: {e}")
            return False


# Factory function to get the appropriate model
def get_model(task_id):
    """
    Get the appropriate GAN model based on task ID
    
    Args:
        task_id (str): Task identifier
        
    Returns:
        BaseGANModel: The appropriate GAN model instance
    """
    models = {
        "task1": StyleTransferGAN(),
        "task2": PhotoEnhancementGAN()
    }
    
    return models.get(task_id)