backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
python-dotenv==1.0.0